Added automatic tabs loading in modules from other modules. To include tabs from offering modules, include this after you've created a tab pannel $tabBox: foreach($all_tabs as $name => $tab)         $tabBox->add($tab, $name);
the $all_tabs array is created in index.php and stores all tabs for other modules. Actually, it can be used for the current module as well.

The tabs that are automatically inserted have names as follows:
$module_to_be_inserted_in . '_tab.anyname_you_like_for_a_title.php';

For an example, look the bottom of the modules/tasks/view.php file and the modules/files/tasks_tab.files.php files.


git-svn-id: http://svn.code.sf.net/p/dotproject/code/trunk@2207 c625438d-3aaf-4c84-a37d-fcf3134530f0