Fix for Mantis Report #2615:
Both getAllowedSQL() and and setAllowedSQL in classes/dp.class.php needed
to be extended so that an alternate permissions module name could be
specified instead of assuming that the calling object's table name and
permission module name were the same and that access could be chaeck using
the table name. This assumption fails in the smartsearch module as it does
not extend the class in classes/dp.class.php and the defined table is not
synonymous with the module which permissions should be checked against. The
smartsearch module now makes use of this extended functionality to correct
the problem.

Also did some clean-up of the smartsearch module to make it more maintainable.

git-svn-id: http://svn.code.sf.net/p/dotproject/code/branches/stable_2@5869 c625438d-3aaf-4c84-a37d-fcf3134530f0