Fixes for Mantis Reports #2486 and #2545:
SMTPSend() in classes/libmail.class.php now tries to connect first using the 
ESMTP keyword "EHLO" and then falls back to older SMTP "HELO". 

Also, since ESMTP functions are potentially available, capability to apply TLS 
to a connection has been added should it be desired or required. Use of TLS is 
toggled by a new config value. If TLS is selected, the fallback to "HELO" is 
not done as the keyword "STARTTLS" is only available under ESMTP.



git-svn-id: http://svn.code.sf.net/p/dotproject/code/branches/stable_2@5880 c625438d-3aaf-4c84-a37d-fcf3134530f0