Allow task duration of a new task to be set to 0 instead of turning empty values into 1. This way you can see which tasks do not have a time estimate yet.
When neither the parent task nor project have an end date, set the default end date of a new task to 1 week after the start date (which defaults to the current date already). This removes the need to select an arbitrary date when entering new tasks with no explicit end date.